+++ CHANGELOG.md
@@ -3,7 +3,10 @@ None
 
 ## New API
 * [`<Client>.on('armSwing', isMainHand => ...)`](https://oscarnow.github.io/minecraft-server/1.1.0/classes/Client#on)
-* [`<Client>.tabItem(tabItemOptions)`](https://oscarnow.github.io/minecraft-server/1.1.0/classes/Client#tabitem)
+
+* [`<Client>.tabItem(tabItemOptions)`](https://oscarnow.github.io/minecraft-server/1.1.0/classes/Client#tabItem)
+* [`<Client>.tabItems`](https://oscarnow.github.io/minecraft-server/1.1.0/classes/Client#tabItems)
+* [`<Client>.on('change', 'tabItems', tabItems => ...)`](https://oscarnow.github.io/minecraft-server/1.1.0/classes/Client#on)
 
 ## Issues fixed
 None